requirements.txt
Click==7.0
etcd3==0.11.1
PyYAML==5.1.2
python-json-logger==0.1.11
requests==2.23.0
parse==1.12.1
pyinotify==0.9.6




